requirements.txt
boto3==1.29.6
botocore==1.32.6
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
etelemetry==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
future==0.18.3
httplib2==0.22.0
isodate==0.6.1
jmespath==1.0.1
jsonpointer==2.1
looseversion==1.3.0
lxml==4.9.3
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.2
pandas==2.1.3
pathlib==1.0.1
prov==2.0.0
pydot==1.4.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
s3transfer==0.7.0
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
